feat: less clutter in output
From now on, regular strings are prefixed with "<", rest is either
non-prefixed base type, e.g., `Infinity` or `NaN`, or short-prefixed
data, e.g., `@one.two` or `[42]`.

I noticed that most of the data is non-string, so no point in adding
both pre- and post-fixes on everything else.
This way logs are shorter and still readable (albeit a bit less, because
"circular" gives meaning straight away, while "@@" does not).